refactor: extract mcpTransportConfig from shared pakage
